postprocessing-service/requirements.txt
# Message Queue
pika>=1.3

# Data handling & Array Operations
numpy>=1.23
pandas>=1.5 # Keep if you plan structured data analysis later

# SORT Algorithm Dependencies
scipy>=1.7        # For linear_sum_assignment (Hungarian algorithm)
filterpy>=1.4.5   # For Kalman Filter implementation

# AWS SDK for Python (DynamoDB)
boto3>=1.26

# Utilities & Config
python-dotenv>=1.0.0 # For loading environment variables

# --- Optional ---
# DB clients (choose one or more based on your storage needs later)
# boto3>=1.26        # for AWS S3 & DynamoDB

# JSON & validation (Keep if needed later)
# jsonschema>=4.17